Updated the version number for appflight.
Added a simple graphical framerate indicator so that the resulting screenshots look a little more like the flash version.

Updated the player entity so that each has it's own instance data. Note to self: stop being bloody lazy when creating classes - always use instance data.

Updated the target entity to use instance data.

Added 100 players instead of 1 to test the frame rate.

Modified the director so that it calculates the framerate, but doesn't draw it.